Remove inconsistent gaiacli keys new/add dualism
- Incorporate --recover in the new command.
- Code clean up

Closes: #2595
